% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_imputation_approaches.R
\name{binomial_draw_gt_imputation}
\alias{binomial_draw_gt_imputation}
\title{Perform GT Imputation, Loci-by-Loci}
\usage{
binomial_draw_gt_imputation(vcfRobj, seed = 48)
}
\arguments{
\item{seed}{numeric; Reproducible Seed}

\item{vcfRobject}{vcfR; a vcfR object}
}
\value{
matrix of GT (0,1,2) with missing loci imputed based on
}
\description{
Perform GT Imputation, Loci-by-Loci
}
\details{
Loci must all be biallelic. Imputation for missing loci is performed by drawing a 0 or 2 weighted by the PLAF
}
