% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{bivcfR2segsites_wsaf}
\alias{bivcfR2segsites_wsaf}
\title{bivcfR2segsites_wsaf}
\usage{
bivcfR2segsites_wsaf(vcfRobject = NULL, err = 0.025)
}
\arguments{
\item{vcfRobject}{vcfR; a biallelic vcfR object}

\item{err}{numeric; the difference between the max WSAF and min WSAF among samples being compared that is tolerated as not being a segregating site.}
}
\description{
Read biallelic vcfR object to segregated sites based on wsaf
}
