% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_filt.R
\name{corMat}
\alias{corMat}
\title{cormat}
\usage{
corMat(m)
}
\description{
Calculates genetic autocorrelation among variants.
}
\details{
Expects a matrix of allele frequencies with rows as loci and columns as samples
}
\author{
Bob Verity & Nick Brazeau
}
