% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_imputation_approaches.R
\name{EB_for_wsaf}
\alias{EB_for_wsaf}
\title{Perform Empirical Bayes Shrinkage, Loci-by-Loci}
\usage{
EB_for_wsaf(vcfRobj, imputemissing = T, seed = 48)
}
\arguments{
\item{imputemissing}{boolean; Should imputation for missing loci be performed}

\item{vcfRobject}{vcfR; a vcfR object}
}
\value{
matrix of wsaf (non-referent alelle) shrunk by a beta-binomial empirical bayes model
}
\description{
Perform Empirical Bayes Shrinkage, Loci-by-Loci
}
\details{
Loci must all be biallelic. Imputation for missing loci is performed by drawing a random number from the beta-binomial model

Note, optim will fail
}
