% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_filt.R
\name{genautocorr}
\alias{genautocorr}
\title{genautocorr setup}
\usage{
genautocorr(vcffile = NULL, vcfR = NULL, biallelicsnps = TRUE)
}
\arguments{
\item{vcffile}{A variant call file (vcf) path. This VCF will be converted to an object of class \code{vcfR}.}
}
\description{
Calculates genetic autocorrelation for later linkage disequilibrium filtering.
}
\details{
From an object of class \code{vcfR}, calculate the genetic autocorrelation as the estimate of linkage disequilibrium by genetic distance.
}
\author{
Bob Verity & Nick Brazeau
}
