% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dab_genet_dist.R
\name{NRAFdf2Fw_pairwisegendist}
\alias{NRAFdf2Fw_pairwisegendist}
\title{NRAFlist2Fw_pairwisegendist}
\usage{
NRAFdf2Fw_pairwisegendist(NRAFdf = NULL, window = 50)
}
\description{
NRAFlist2Fw_pairwisegendist
}
\author{
Nick Brazeau
}
