% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpl_slctn.R
\name{select_samples}
\alias{select_samples}
\title{select_samples}
\usage{
select_samples(vcfRobject = NULL, smplvctr = NULL)
}
\arguments{
\item{smplvctr}{vector; A character vector that corresponds to the samples that should be extracted}
}
\description{
Select samples from a \code{vcfR}
}
