% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loci_selection.R
\name{vcf_loci_splitter}
\alias{vcf_loci_splitter}
\title{Split a VCF to Individual Loci}
\usage{
vcf_loci_splitter(vcfRobj, write_new_meta = F)
}
\arguments{
\item{vcfRobj}{S4object; vcfR object from the vcfR package}

\item{write_new_meta}{boolean; Flag for whether or not the original vcf metadata should be included.}
}
\value{
List of vcfRobj
}
\description{
Split a VCF to Individual Loci
}
\details{
If write_new_meta is set to False, a simplified vcf header will be returned (just file format)
}
