% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{vcffilter_ChromPos}
\alias{vcffilter_ChromPos}
\title{vcffilter_ChromPos}
\usage{
vcffilter_ChromPos(vcfRobject = NULL, chromposbed = NULL)
}
\arguments{
\item{chromposdf}{an dataframe that contains loci to be excluded and is in the form of a bedfile or has
been produced by \code{GFF2VariantAnnotation_Short}}
}
\description{
Produces a subsetted \code{vcfR} with specific loci excluded as determined \code{chromposdf}.
}
\details{
chromposdf must be a tibble with columns: seqname (where chromomsome information is stored), start, end, geneid. The logic
behding the "seqname" is allow users to carry around chromosome names that may be different in the bed file and the vcf
}
