% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO_Format_filt.R
\name{vcffilter_format}
\alias{vcffilter_format}
\title{vcfR2vcffilter_FORMAT}
\usage{
vcffilter_format(vcfRobject = NULL, formatGQ = NULL, formatDP = NULL,
  formatSP = NULL, prop.loci.missing = NULL, biallelic = TRUE,
  SNPs = TRUE)
}
\arguments{
\item{prop.loci.missing}{Given a loci, how many samples can have missing information before that loci is dropped}
}
\description{
vcfR2vcffilter_FORMAT
}
