% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO_Format_filt.R
\name{vcffilter_info}
\alias{vcffilter_info}
\title{vcfR2vcffilter_INFO}
\usage{
vcffilter_info(
  vcfRobject = NULL,
  infoMQ = NULL,
  infoQD = NULL,
  infoSOR = NULL,
  infoAF = NULL,
  infoDP = NULL,
  infoFS = NULL,
  infoMQRankSum = NULL,
  infoReadPosRankSum = NULL,
  biallelic = TRUE,
  SNPs = TRUE
)
}
\arguments{
\item{infoDP}{options is based on a percentile cutoff. As a result, if you specify 0.1, the top and bottom 10th read depth percentiles will be excluded}
}
\description{
vcfR2vcffilter_INFO
}
