% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dab_genet_dist.R
\name{vcfR2Fw_pairwisegendist}
\alias{vcfR2Fw_pairwisegendist}
\title{vcfR2Fw_pairwisegendist}
\usage{
vcfR2Fw_pairwisegendist(vcfRobject = NULL, biallelicsnps = TRUE,
  segsites = 0.001, window = 50)
}
\description{
vcfR2Fw_pairwisegendist
}
\author{
Nick Brazeau
}
