% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{vcfR2nohetsmonoclonal_gt}
\alias{vcfR2nohetsmonoclonal_gt}
\title{Remove heterozygote calls and make monoclonal}
\usage{
vcfR2nohetsmonoclonal_gt(vcfRobj)
}
\description{
Read vcfR object and code all het sites in GT as missing
}
