% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{vcfR2removepoorcovloci_dp}
\alias{vcfR2removepoorcovloci_dp}
\title{vcfR2removepoorcovloci_dp}
\usage{
vcfR2removepoorcovloci_dp(vcfRobject = NULL,
  minsampleswithcoverage = 1)
}
\arguments{
\item{vcfRobject}{vcfR; a vcfR object}

\item{minsampleswithcoverage}{numeric; the minimum number of samples with coverage \(i.e. DP > 0) needed for a loci to be kept}
}
\description{
Read vcfR object to remove loci where too few samples have coverage
}
