% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{vcfR2removesingletons_gt}
\alias{vcfR2removesingletons_gt}
\title{vcfR2removesingletons_gt}
\usage{
vcfR2removesingletons_gt(vcfRobj)
}
\description{
Read vcfR object and remove all loci that have are singletons
}
