% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loci_selection.R
\name{vcfR2SubsetChrom}
\alias{vcfR2SubsetChrom}
\title{vcfR2SubsetChrom}
\usage{
vcfR2SubsetChrom(vcfRobject = NULL, chromvect = NULL)
}
\description{
Produces a subsetted \code{vcfR} based on chromosome
}
