% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{vcfR2SubsetChromPos}
\alias{vcfR2SubsetChromPos}
\title{vcfR2SubsetChromPos}
\usage{
vcfR2SubsetChromPos(vcfRobject = NULL, chromposbed = NULL)
}
\description{
Produces a subsetted \code{vcfR} based on chromosome. Note the chromosome name must be contained in the column \code{seqname}
}
