% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dab_genet_dist.R
\name{wicalc}
\alias{wicalc}
\title{DAB weight calculator}
\usage{
wicalc(NRAFdf, window = 50)
}
\description{
This is based on the within sample Fw gen dist function described in PMID: 26943619
}
\details{
Inspired by Bob Verity's corMat function
}
\author{
Nick Brazeau
}
