% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_filt.R
\name{calc_loci_missingness_by_smpl}
\alias{calc_loci_missingness_by_smpl}
\title{Filter VCF based on given loci missingness by sample}
\usage{
calc_loci_missingness_by_smpl(vcfRobject = NULL)
}
\arguments{
\item{vcfRobject}{S4 object; A vcf that has been converted to a \code{vcfR} object}
}
\description{
Calculate the proportion of the loci that are missing by samples from a \code{vcfR} object
}
