% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loci_selection.R
\name{select_samples}
\alias{select_samples}
\title{Select Samples from VCF}
\usage{
select_samples(vcfRobject = NULL, smplvctr = NULL)
}
\arguments{
\item{smplvctr}{vector; A character vector that corresponds to the samples that should be extracted}
}
\description{
Select samples from a \code{vcfR}
}
