% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_filt.R
\name{vcfR2LDfiltered}
\alias{vcfR2LDfiltered}
\title{vcfR2LDfiltered}
\usage{
vcfR2LDfiltered(
  vcffile = NULL,
  vcfR = NULL,
  genautocorrresult = NULL,
  threshDist = NULL,
  threshR2 = NULL,
  biallelicsnps = TRUE
)
}
\arguments{
\item{vcffile}{A variant call file (vcf) path. This VCF will be converted to an object of class \code{vcfR}.}

\item{threshDist}{Genetic distance to filter at.}

\item{threshR2}{Genetic autocorrelation to filter at}
}
\description{
Filtering an object of class \code{vcfR} for linkage disequilibrium via genetic autocorrelation.
}
